work version 1.2

+++ src/main/java/ru/bureau/settings/dto/SettingDto.java
@@ -4,22 +4,22 @@ package ru.bureau.settings.dto;
 import java.io.Serializable;
 
 public class SettingDto implements Serializable {
-    private Long id;
+    private Integer id;
     private String name;
     private String value;
 
     // Конструкторы
     public SettingDto() {}
 
-    public SettingDto(Long id, String name, String value) {
+    public SettingDto(Integer id, String name, String value) {
         this.id = id;
         this.name = name;
         this.value = value;
     }
 
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
     public String getValue() { return value; }
